+++ src/api/serializers/transactions.ts
@@ -9,25 +9,9 @@ import {
   TransactionSummary,
   TransactionStatus,
 } from '../schemas/v3/entities/transaction-summaries.js';
-import {
-  DbPrincipalTransactionSummary,
-  DbTransaction,
-  DbTransactionSummary,
-} from '../../datastore/v3/types.js';
+import { DbPrincipalTransactionSummary, DbTransactionSummary } from '../../datastore/v3/types.js';
 import { DbTxStatus, DbTxTypeId } from '../../datastore/common.js';
 import { getTxTenureChangeCauseString } from '../controllers/db-controller.js';
-import {
-  BaseTransaction,
-  CoinbaseTransaction,
-  ContractCallTransaction,
-  PoisonMicroblockTransaction,
-  SmartContractTransaction,
-  TenureChangeTransaction,
-  TokenTransferTransaction,
-  Transaction,
-} from '../schemas/v3/entities/transactions.js';
-import codec from '@stacks/codec';
-import { serializePostCondition } from './post-conditions.js';
 import { PrincipalTransactionSummary } from '../schemas/v3/entities/principal-transactions.js';
 
 /**
@@ -146,113 +130,6 @@ export function parseDbTransactionSummary(summary: DbTransactionSummary): Transa
     }
     default:
       throw new Error(`Unexpected DbTxTypeId: ${summary.type_id}`);
-  }
-}
-
-/**
- * Parses a database transaction into a transaction.
- * @param transaction - The database transaction to parse.
- * @returns The parsed transaction.
- */
-export function parseDbTransaction(transaction: DbTransaction): Transaction {
-  const summary = parseDbTransactionSummary(transaction);
-  const decodedPostConditions = codec.decodePostConditions(transaction.post_conditions);
-  const result: BaseTransaction = {
-    ...summary,
-    parent_block: {
-      hash: transaction.parent_block_hash,
-      index_hash: transaction.parent_index_block_hash,
-    },
-    post_conditions: decodedPostConditions.post_conditions.map(pc => serializePostCondition(pc)),
-    event_count: transaction.event_count,
-    execution_cost: {
-      read_count: transaction.execution_cost_read_count,
-      read_length: transaction.execution_cost_read_length,
-      runtime: transaction.execution_cost_runtime,
-      write_count: transaction.execution_cost_write_count,
-      write_length: transaction.execution_cost_write_length,
-    },
-    vm_error: transaction.vm_error,
-  };
-  switch (transaction.type_id) {
-    case DbTxTypeId.TokenTransfer: {
-      const tokenTransfer: TokenTransferTransaction = {
-        ...result,
-        type: 'token_transfer',
-        token_transfer: {
-          recipient: transaction.token_transfer_recipient_address!,
-          amount: transaction.token_transfer_amount!,
-          memo: transaction.token_transfer_memo,
-        },
-      };
-      return tokenTransfer;
-    }
-    case DbTxTypeId.SmartContract: {
-      const smartContract: SmartContractTransaction = {
-        ...result,
-        type: 'smart_contract',
-        smart_contract: {
-          clarity_version: transaction.smart_contract_clarity_version,
-          contract_id: transaction.smart_contract_contract_id!,
-          source_code: transaction.smart_contract_source_code!,
-        },
-      };
-      return smartContract;
-    }
-    case DbTxTypeId.ContractCall: {
-      const contractCall: ContractCallTransaction = {
-        ...result,
-        type: 'contract_call',
-        contract_call: {
-          contract_id: transaction.contract_call_contract_id!,
-          function_name: transaction.contract_call_function_name!,
-          function_args: codec
-            .decodeClarityValueList(transaction.contract_call_function_args!)
-            .map(c => ({
-              hex: c.hex,
-              repr: c.repr,
-            })),
-        },
-      };
-      return contractCall;
-    }
-    case DbTxTypeId.PoisonMicroblock: {
-      const poisonMicroblock: PoisonMicroblockTransaction = {
-        ...result,
-        type: 'poison_microblock',
-      };
-      return poisonMicroblock;
-    }
-    case DbTxTypeId.Coinbase: {
-      const coinbase: CoinbaseTransaction = {
-        ...result,
-        type: 'coinbase',
-        coinbase: {
-          alt_recipient: transaction.coinbase_alt_recipient,
-          payload: transaction.coinbase_payload!,
-          vrf_proof: transaction.coinbase_vrf_proof,
-        },
-      };
-      return coinbase;
-    }
-    case DbTxTypeId.TenureChange: {
-      const tenureChange: TenureChangeTransaction = {
-        ...result,
-        type: 'tenure_change',
-        tenure_change: {
-          cause: getTxTenureChangeCauseString(transaction.tenure_change_cause!),
-          tenure_consensus_hash: transaction.tenure_change_tenure_consensus_hash!,
-          prev_tenure_consensus_hash: transaction.tenure_change_prev_tenure_consensus_hash!,
-          burn_view_consensus_hash: transaction.tenure_change_burn_view_consensus_hash!,
-          previous_tenure_end: transaction.tenure_change_previous_tenure_end!,
-          previous_tenure_blocks: transaction.tenure_change_previous_tenure_blocks!,
-          pubkey_hash: transaction.tenure_change_pubkey_hash!,
-        },
-      };
-      return tenureChange;
-    }
-    default:
-      throw new Error(`Unexpected DbTxTypeId: ${transaction.type_id}`);
   }
 }
